comment section partitioning, prettier, and index key
