+++ .eslintrc.json
@@ -1,6 +1,3 @@
 {
-  "extends": [
-    "next/core-web-vitals",
-    "next/typescript"
-  ]
+  "extends": ["next/core-web-vitals", "next/typescript"]
 }
