+++ src/app/layout.tsx
@@ -1,15 +1,14 @@
 import "tailwind.config";
 import { Metadata } from "next";
 import Footer from "../ui/footer/footer";
-import Header from "../ui/header/index"
-import './global.css';
-import Head from 'next/head';
+import Header from "../ui/header/index";
+import "./global.css";
+import Head from "next/head";
 
 export const metadata: Metadata = {
   title: "Unified Portal",
   description: "The tryout for the Unified Portal.",
 };
-
 
 export default function RootLayout({
   children,
@@ -18,10 +17,10 @@ export default function RootLayout({
 }) {
   return (
     <html lang="en">
-       <Head>
+      <Head>
         <link rel="icon" href="/favicon.ico" />
       </Head>
-      <body className="bg-Zwhite">
+      <body className="bg-rm-white">
         <Header />
         {children}
         <Footer />
