+++ src/ui/comment/comments.json
@@ -1,0 +1,17 @@
+[
+  {
+    "name": "أحمد عبدالله",
+    "time": "20/10/2024",
+    "comment": "الاستعلام عن معاملة الكترونية وصف الخدمة. يقوم المستفيد بالاستعلام عن المعاملة التي تم إنشاءها من قبله وآخر عملية تمت عليها من الاتصالات الادارية · اخر تحديث"
+  },
+  {
+    "name": "مجد كارم",
+    "time": "20/10/2024",
+    "comment": "الاستعلام عن معاملة الكترونية وصف الخدمة. يقوم المستفيد بالاستعلام عن المعاملة التي تم إنشاءها من قبله وآخر عملية تمت عليها من الاتصالات الادارية · اخر تحديث"
+  },
+  {
+    "name": "ولاء عبدالعزيز",
+    "time": "20/10/2024",
+    "comment": "الاستعلام عن معاملة الكترونية وصف الخدمة. يقوم المستفيد بالاستعلام عن المعاملة التي تم إنشاءها من قبله وآخر عملية تمت عليها من الاتصالات الادارية · اخر تحديث"
+  }
+]
