+++ src/ui/elements/breadcrumb.tsx
@@ -14,8 +14,8 @@ export default function Breadcrumb({ links }: BreadcrumbProps) {
     <div className=" ms-4 me-4 mb-5 rounded-b-3xl bg-[#d7e5de] label12 flex items-center justify-end px-4 py-2 lg:px-6 lg:py-2">
       <nav aria-label="breadcrumb">
         <ol className="flex flex-row-reverse items-center gap-2.5 lg:gap-1.5">
-          {links.map((link) => (
-            <li key={link.href} className="gap-1.5">
+          {links.map((link, index) => (
+            <li key={index} className="gap-1.5">
               {link.element}
             </li>
           ))}
