+++ src/ui/footer/footer-list-lg.tsx
@@ -1,12 +1,12 @@
 import FooterUl from "./footer-ul";
-import Links from "./Links.json"
+import Links from "./Links.json";
 
 export default function FooterListLg() {
   return (
     <div className="lg:flex hidden justify-end gap-4 pb-4 px-2 w-3/4">
-        <FooterUl {...Links.important_links} />
-        <FooterUl {...Links.support_links} />
-        <FooterUl {...Links.overview_links} />
+      <FooterUl {...Links.important_links} />
+      <FooterUl {...Links.support_links} />
+      <FooterUl {...Links.overview_links} />
     </div>
   );
 }
