+++ src/ui/header/Links.json
@@ -1,76 +1,76 @@
 {
-    "aboutAmana": {
-        "name": "عن الأمانة",
-        "links": [
-            {
-                "href": "##",
-                "label": "عن الأمانة"
-            },
-            {
-                "href": "#",
-                "label": "عن مدينة الرياض"
-            },
-            {
-                "href": "####",
-                "label": "أمناء الرياض"
-            },
-            {
-                "href": "###",
-                "label": "البلديات"
-            },
-            {
-                "href": "######",
-                "label": "شركاء الأمانة المحليين"
-            },
-            {
-                "href": "#####",
-                "label": "المبادرات و المشاريع"
-            },
-            {
-                "href": "########",
-                "label": "التنمية المستدامة"
-            },
-            {
-                "href": "#######",
-                "label": "إحصائيات البوابة"
-            },
-            {
-                "href": "##########",
-                "label": "الأمانة في أرقام"
-            },
-            {
-                "href": "#########",
-                "label": "الجوائز"
-            }
-        ]
-    },
-    "openDataSet": {
-        "name": "البيانات المفتوحة",
-        "links": [
-            {
-                "href": "#",
-                "label": "البيانات المفتوحة"
-            },
-            {
-                "href": "##",
-                "label": "البيانات الحكومية المفتوحة"
-            },
-            {
-                "href": "###",
-                "label": "مجموعات  البيانات"
-            },
-            {
-                "href": "####",
-                "label": "البوابة الجيومكانية"
-            },
-            {
-                "href": "#####",
-                "label": "طلب اضافة مفوض"
-            },
-            {
-                "href": "######",
-                "label": "قصص نجاح البيانات المفتوحة"
-            }
-        ]
-    }
+  "aboutAmana": {
+    "name": "عن الأمانة",
+    "links": [
+      {
+        "href": "##",
+        "label": "عن الأمانة"
+      },
+      {
+        "href": "#",
+        "label": "عن مدينة الرياض"
+      },
+      {
+        "href": "####",
+        "label": "أمناء الرياض"
+      },
+      {
+        "href": "###",
+        "label": "البلديات"
+      },
+      {
+        "href": "######",
+        "label": "شركاء الأمانة المحليين"
+      },
+      {
+        "href": "#####",
+        "label": "المبادرات و المشاريع"
+      },
+      {
+        "href": "########",
+        "label": "التنمية المستدامة"
+      },
+      {
+        "href": "#######",
+        "label": "إحصائيات البوابة"
+      },
+      {
+        "href": "##########",
+        "label": "الأمانة في أرقام"
+      },
+      {
+        "href": "#########",
+        "label": "الجوائز"
+      }
+    ]
+  },
+  "openDataSet": {
+    "name": "البيانات المفتوحة",
+    "links": [
+      {
+        "href": "#",
+        "label": "البيانات المفتوحة"
+      },
+      {
+        "href": "##",
+        "label": "البيانات الحكومية المفتوحة"
+      },
+      {
+        "href": "###",
+        "label": "مجموعات  البيانات"
+      },
+      {
+        "href": "####",
+        "label": "البوابة الجيومكانية"
+      },
+      {
+        "href": "#####",
+        "label": "طلب اضافة مفوض"
+      },
+      {
+        "href": "######",
+        "label": "قصص نجاح البيانات المفتوحة"
+      }
+    ]
+  }
 }
