+++ src/ui/header/index.tsx
@@ -1,11 +1,11 @@
 import NavBarLG from "./nav-bar-lg";
 import NavBarMS from "./nav-bar-ms";
 
-export default function Header(){
-    return (
-        <>
-        <NavBarLG />
-        <NavBarMS />
-        </>
-    );
+export default function Header() {
+  return (
+    <>
+      <NavBarLG />
+      <NavBarMS />
+    </>
+  );
 }
